+++ src/zts-helpers.ts
@@ -1,4 +1,4 @@
-import ts from "typescript";
+import type ts from "typescript";
 import { z } from "zod";
 import { FlatObject } from "./common-helpers";
 import { SchemaHandler } from "./schema-walker";
@@ -15,20 +15,3 @@ export interface ZTSContext extends FlatObject {
 }
 
 export type Producer = SchemaHandler<ts.TypeNode, ZTSContext>;
-
-const primitives: ts.KeywordTypeSyntaxKind[] = [
-  ts.SyntaxKind.AnyKeyword,
-  ts.SyntaxKind.BigIntKeyword,
-  ts.SyntaxKind.BooleanKeyword,
-  ts.SyntaxKind.NeverKeyword,
-  ts.SyntaxKind.NumberKeyword,
-  ts.SyntaxKind.ObjectKeyword,
-  ts.SyntaxKind.StringKeyword,
-  ts.SyntaxKind.SymbolKeyword,
-  ts.SyntaxKind.UndefinedKeyword,
-  ts.SyntaxKind.UnknownKeyword,
-  ts.SyntaxKind.VoidKeyword,
-];
-
-export const isPrimitive = (node: ts.TypeNode): node is ts.KeywordTypeNode =>
-  (primitives as ts.SyntaxKind[]).includes(node.kind);
